pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "perfect-game"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "imagekitio>=5.4.0",
    "python-multipart>=0.0.6",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "routers*"]
